requirements.txt
asgiref==3.7.2
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
